move usage_key_with_run function
