+++ cms/djangoapps/contentstore/xblock_storage_handlers/__init__.py
@@ -2,5 +2,5 @@
 Xblock services that contain the business logic for xblock views.
 """
 from .create_xblock import *
-from .xblock_helpers import *
 from .xblock_service import *
+from .usage_key_with_run import *
