+++ cms/djangoapps/contentstore/xblock_storage_handlers/create_xblock.py
@@ -11,7 +11,7 @@ from xmodule.tabs import StaticTab
 from cms.djangoapps.models.settings.course_grading import CourseGradingModel
 from openedx.core.toggles import ENTRANCE_EXAMS
 
-from .xblock_helpers import usage_key_with_run
+from .usage_key_with_run import usage_key_with_run
 from ..helpers import GRADER_TYPES, remove_entrance_exam_graders
